Make minor code, comment, and formatting changes
- Make minor code, comment, and formatting changes throughout the code base
- Add Bash completion script to the Docker configuration file [`Dockerfile`]
- Add profile rule to Make configurations for gprof support
- Update the README file, conformance clause, and language guide
- Fix parsing bugs in the Bash completion script
- Add comments to various configuration scripts, providing usage examples
- Fix various MDC error code bugs and inconsistencies
- Fix potential double free bug in the interpreter
- Refactor the MATH utility library into five routines instead of seven
- Fix bug in `^%D` that prevents display of the first day of `$HOROLOG`
- Fix bug in `^%ZRCHG` that calls the wrong routine name for recompiling
- Fix bug in `^%ZRSE` that executes code out of global nodes that don't exist

+++ include/compile.h
@@ -50,7 +50,7 @@
 }                                                                               // report an error
 
 #define INDSNOK(size) (((size * 2) + (sizeof(int) * 2) + isp) > MAX_ISTK)       // For testing indirection size - a guess
-#define INDANOK(addr) ((addr + (sizeof(int) * 2) + 1) >= &indstk[MAX_ISTK])     // For testing the address of compiled indirection
+#define INDANOK(addr) ((addr + (sizeof(long) * 2) + 1) >= &indstk[MAX_ISTK])    // For testing the address of compiled indirection
 
 #define RBD_OVERHEAD (sizeof(rbd *) + (sizeof(u_int) * 2) + sizeof(time_t) \
                      + sizeof(var_u) + (sizeof(u_char) * 2) + sizeof(u_short))
